support overwrite file

+++ upload.php
@@ -1,5 +1,9 @@
 <?php
 
+/**
+ * to upload file at jsondata directory
+ */
+
 if (isset($_POST['upload'])) {
     //var_dump($_FILES['file']);
     echo count($_FILES['file']['name']);
@@ -7,10 +11,16 @@ if (isset($_POST['upload'])) {
     for ($i=0; $i<count($_FILES['file']['name']);$i++) {
         echo $_FILES['file']['name'][$i];
         echo "<br />";
+        print_r($_FILES['file']['name'][$i]);
         $file_name= $_FILES['file']['name'][$i];
         $tmp_name = $_FILES['file']['tmp_name'][$i];
-        move_uploaded_file($tmp_name, 'jsondata/'.$file_name);
+
+        if (file_exists("jsondata/".$file_name)) {
+            unlink("jsondata".$file_name);
+            move_uploaded_file($tmp_name, 'jsondata/'.$file_name);
+        } else {
+            move_uploaded_file($tmp_name, 'jsondata/'.$file_name);
+        }
     }
     header('Location: index.html');
 }
-?>
